+++ test/functional/uuid/uuid.ts
@@ -22,7 +22,7 @@ describe("uuid type", () => {
 
     it("should make correct schema with Postgres' uuid type", () => Promise.all(connections.map(async connection => {
         await connection.syncSchema(true);
-        const queryRunner = connection.driver.createQueryRunner();
+        const queryRunner = connection.createQueryRunner();
         let schema = await queryRunner.loadTableSchema("record");
         await queryRunner.release();
         expect(schema).not.to.be.empty;